RB tree tests succeed. Potential optimizations exist.
The issue resolved was, ultimately, that there were a couple edge cases that weren't covered in the documentation I was reading that only appear when nil nodes are actually nil, and do not know about their parents. What is effectively a fuzz test, rb-test suceeds at inserting and deleting 10000 random elements from the rb tree.